Move test around
